Move shortcut reconciliation into the setting registry.
